sentinel_code/backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
langgraph>=0.0.10
langchain>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0
pytest>=7.4.0
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
